started with refactor of prediction intervals for forecasters
